+++ code/appdata/curves_test.go
@@ -16,6 +16,6 @@ func TestSomething(t *testing.T) {
 	curve_data := LoadCurveData(appConfig, "inc_smooth")
 
 	assert.NotNil(t, curve_data, "Curve data found, loaded, not nil")
-	assert.Equal(t, curve_data.Values[0], float32(0), "First value of this curve should be 0")
+	assert.Equal(t, curve_data.Values[0], float64(0), "First value of this curve should be 0")
 
 }
